Rework of face unwraping done. Polishing ahead, figure out what is 'EXC_I386_GPFLT' err

+++ libft/ft_strcpy.c
@@ -14,7 +14,7 @@
 
 char	*ft_strcpy(char *dst, char *src)
 {
-	ssize_t	b;
+	long long	b;
 
 	b = -1;
 	while (src[++b] != '\0')
